Rename scripts to pep_patch_hydrophobic and pep_patch_electrostatic
